Got it to save messages and display them again.
Currently it pulls the entire list from Firebase each time, which is inefficient. We should update it to only seek new messages. I think we can do that by storing the old messages on the phone and then only checking for new children, but we'll have to play around with it and see...